refactor(nextwiki): 统一使用 MapStruct 进行对象转换
- 将 NextwikiConverter 替换为 NextwikiStructMapper，简化代码并提高一致性
- 更新多个服务中 VO 与 DTO 之间的转换调用
- 统一文件权限控制、文件版本处理、文件夹拷贝等模块的转换实现

fix(agent): 增加事务回滚异常配置

- 在 AgentDefinitionServiceImpl 的 create、update、removeById 方法中添加 rollbackFor = Exception.class
- 确保事务在发生异常时能够正确回滚，提升数据一致性安全

perf(agent): 设定批量对话异步调用超时限制

- 在 AgentFacadeImpl 中添加批量总超时常量 BATCH_CHAT_TOTAL_TIMEOUT_SECONDS = 60秒
- 使用 CompletableFuture 的 orTimeout 方法防止单任务挂起影响整个批量执行

chore(config): 支持生成器模块数据库配置通过环境变量注入

- ydsz-generator-app 与 ydsz-generator-web 模块的 datasource 配置增加对 NACOS_GEN_DB_* 环境变量的支持
- 使配置更加灵活且易于 CI/CD 管理

fix(common-exception): 增加异常级别支持及传递

- 在 BaseExceptionHandler 中设置异常信息的 level 属性
- 在构造标准错误响应时支持异常级别参数，丰富错误事件数据
- 异常响应结构新增 level 字段，用于前端错误级别展示和处理

feat(nextwiki): 优化摘要和相似度算法文档注释

- 补充 generateSummaryByTextRank 方法的算法说明，强调使用 primitive double 类型
- 详细描述 calculateSimilarity 方法的算法与类型选择原因，提高代码可读性

refactor(nextwiki): 使用 BigDecimal 精确计算批量进度百分比

- 在 BatchProgressNotifier 中替换 double 直接计算为 BigDecimal 运算
- 遵循行业规范，避免浮点精度问题，确保进度显示准确性

feat(gateway): 新增 SQL 注入检测支持

- 在 bootstrap.yml 配置中添加 gateway.sql-injection 相关配置项及默认参数
- GatewayFilterConfig 新增 SqlInjectionProperties 支持
- GatewayFilterOrder 中新增 SQL_INJECTION 过滤器顺序定义
- 定义新的错误码 SQL_INJECTION_DETECTED 和 XSS_DETECTED，用以区分安全拦截类型

chore(infra): 优化 MyBatis XML 查询，精准映射数据库字段

- 修改 FileNodeMapper.xml、FileVersionMapper.xml、FileAclMapper.xml、FlowAdminRoleMapper.xml 等查询语句，
  显式指定查询字段，避免使用 SELECT *，提升查询效率与稳定性

style(nextwiki): 优化代码注释及日志

- 统一部分代码注释格式及内容，如 ChunkUploadApplicationService 的 mapper 注释更新
- FileLockController 中更新 VO 转换调用，保持风格一致
- 细节处修正日志代码格式，提高可维护性

+++ ydsz-nextwiki/ydsz-nextwiki-domain/src/main/java/com/njydsz/nextwiki/domain/vo/SearchResultVO.java
@@ -1,6 +1,7 @@
 package com.njydsz.nextwiki.domain.vo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -65,7 +66,7 @@ public class SearchResultVO implements Serializable {
     private String highlight;
 
     @Schema(description = "匹配分数")
-    private Float score;
+    private BigDecimal score;
 
     @Schema(description = "标签列表")
     private List<String> tags;
